backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
python-dotenv==0.19.0
pymongo==4.6.1
google-api-python-client==2.86.0
numpy==1.24.3
scikit-learn==1.2.2
spacy==3.5.3
transformers==4.30.2
torch==2.0.1

# Download spaCy model:
# python -m spacy download en_core_web_sm